Collision Manager class created
Started adding collision detections. Used Spatial HashMap algorithm to determine macro collisions between colliders.
The system allows colliders to be added to a Collision Manager.
The Collision Manager will eventually do the macro and micro collisions and passing on the data to the pObj to resolve the collisions.
Support for boxes in multiple cells is not available.
HASN'T BEEN TESTED

Fixed the delta in update was a slight error
Added debug FPS and fixed error with rotating vertices being put at 0, 0
Added toString function for Vector2

Signed-off-by: obwan02 <obwan02@hotmail.co.nz>